# backend/docs/swagger.yaml
basePath: /api
definitions:
  dto.UnreadCount:
    properties:
      chat_id:
        type: integer
      count:
        type: integer
    type: object
...
  model.FileMetadata:
    properties:
      created_at:
        type: string
      file_size:
        type: integer
      height:
        type: integer
      id:
        type: integer
      message_id:
        type: integer
      mime_type:
        type: string
      original_name:
        type: string
      url:
        type: string
      width:
        type: integer
    type: object
  model.FileMetadataPayload:
    properties:
      file_size:
        type: integer
      height:
        type: integer
      mime_type:
        type: string
      original_name:
        type: string
      url:
        type: string
      width:
        type: integer
    type: object
...
  model.MessageResponse:
    properties:
      chat_id:
        type: integer
      content:
        type: string
      content_type:
        type: string
      created_at:
        type: string
      file_metadata:
        $ref: '#/definitions/model.FileMetadata'
      id:
        type: integer
      quoted_content:
        type: string
      quoted_content_type:
        type: string
      quoted_file_metadata:
        $ref: '#/definitions/model.FileMetadata'
      quoted_message_id:
        type: integer
      quoted_sender_id:
        type: integer
      quoted_sender_name:
        type: string
      sender_id:
        type: integer
      status:
        type: string
    type: object
...
  model.SendMessageRequest:
    properties:
      chat_id:
        type: integer
      content:
        type: string
      content_type:
        type: string
      file_metadata:
        $ref: '#/definitions/model.FileMetadataPayload'
      quoted_message_id:
        type: integer
    required:
    - chat_id
    - content
    type: object
...
  types.Response:
    properties:
      code: {}
      data: {}
      message:
        type: string
    type: object
host: localhost:8080
info:
  contact:
    email: support@wetalk.app
    name: API Support
    url: http://www.wetalk.app/support
  description: WeTalk 在线聊天应用 API 文档
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: http://swagger.io/terms/
  title: WeTalk API
  version: 1.0.0
paths:
...
  /api/messages:
    get:
      description: 获取指定聊天的历史消息记录（按时间倒序，支持分页）
      parameters:
      - description: 聊天 ID
        in: query
        name: chat_id
        required: true
        type: integer
      - description: 偏移量（默认 0）
        in: query
        name: offset
        type: integer
      - description: 每页数量（默认 50）
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 成功
          schema:
            allOf:
            - $ref: '#/definitions/types.Response'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/model.MessageResponse'
                  type: array
              type: object
        "400":
          description: 参数错误
          schema:
            $ref: '#/definitions/types.Response'
        "401":
          description: 未认证
          schema:
            $ref: '#/definitions/types.Response'
      security:
      - BearerAuth: []
      summary: 获取聊天记录
      tags:
      - 消息
    post:
      consumes:
      - application/json
      description: 发送文本/图片/文件消息，支持引用回复；通过 WebSocket 实时推送给聊天中的在线成员
      parameters:
      - description: 消息内容
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/model.SendMessageRequest'
      produces:
      - application/json
      responses:
        "201":
          description: 消息已发送
          schema:
            allOf:
            - $ref: '#/definitions/types.Response'
            - properties:
                data:
                  $ref: '#/definitions/model.MessageResponse'
              type: object
        "400":
          description: 参数错误
          schema:
            $ref: '#/definitions/types.Response'
        "401":
          description: 未认证
          schema:
            $ref: '#/definitions/types.Response'
        "403":
          description: 不是聊天成员
          schema:
            $ref: '#/definitions/types.Response'
      security:
      - BearerAuth: []
      summary: 发送消息
      tags:
      - 消息
  /api/messages/read:
    put:
      consumes:
      - application/json
      description: 将指定聊天中来自对方的消息全部标记为已读
      parameters:
      - description: 聊天 ID
        in: body
        name: request
        required: true
        schema:
          properties:
            chat_id:
              type: integer
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: 已标记已读
          schema:
            $ref: '#/definitions/types.Response'
        "400":
          description: 参数错误
          schema:
            $ref: '#/definitions/types.Response'
        "401":
          description: 未认证
          schema:
            $ref: '#/definitions/types.Response'
      security:
      - BearerAuth: []
      summary: 标记消息已读
      tags:
      - 消息
  /api/messages/unread:
    get:
      description: 获取当前用户在所有聊天中的未读消息数量（供侧边栏角标使用）
      produces:
      - application/json
      responses:
        "200":
          description: 成功
          schema:
            allOf:
            - $ref: '#/definitions/types.Response'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/dto.UnreadCount'
                  type: array
              type: object
        "401":
          description: 未认证
          schema:
            $ref: '#/definitions/types.Response'
      security:
      - BearerAuth: []
      summary: 获取未读消息数
      tags:
      - 消息
...
schemes:
- http
securityDefinitions:
  BearerAuth:
    description: 输入 "Bearer " + JWT token
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"